Tell the composition from what a person is handed
One type was both. Its own words said so in two places — "the rows one
run hands a person" at the top, "not what a person is handed" over the
factory — and `answered` meant "nobody asked" on one and "the rows settle
nothing" on the other. So `GeneratedRows.of(Offering.composed(...))`
compiled, and the reduction this change exists for could be walked past
by writing the shorter of two names.

`Composition` is what the two searches composed. `Offering` is what is
left once what each row would settle has been asked, and the only way to
make one is to ask: `Composition.keeping` takes the rows kept and the
items they answer. A renderer takes an offering and there is no longer a
way to hand it the other thing.

`OfferedRow` moves out to itself, being neither's.

For a caller with no store — a test that builds its fillings by hand, so
there is nothing to run the rows against — `offeringEverything` says
that in a name. Every row goes out and nothing is said to be answered,
which is what not asking means.

And `WhereARowWasAskedFor` no longer restates which points those are.
`worthSearching()` is the authority, and the sentence naming its two
causes was wider than it: a point with no rows to measure is one a
candidate is asked for.
