+++ souther-compiler/src/test/java/souther/compiler/AGenerationThatWentOnDoesNotSayItStoppedTest.java
@@ -84,10 +84,10 @@ class AGenerationThatWentOnDoesNotSayItStoppedTest {
     private static String written(souther.compiler.partition.FillResult result) {
         // The composition and not what is offered: this filling is built here rather than searched
         // for, so there is no store to ask what its rows would settle.
-        return GeneratedRows.of(souther.compiler.query.Offering.composed(
+        return GeneratedRows.of(souther.compiler.query.Composition.composed(
                         souther.compiler.query.OfferingRequest.overTheModule("example.trip", false),
                         Map.of("submit", new Adequacy.Filling(result,
-                                Generator.GenerationResult.NONE, List.of())), null),
+                                Generator.GenerationResult.NONE, List.of())), null).offeringEverything(),
                 Map.of(), SourceNameResolver.identity()).text();
     }
