+++ souther-compiler/src/test/java/souther/compiler/WhatIsWrittenInAnEnsuresIsQuotedOverTheRowsTest.java
@@ -253,9 +253,9 @@ class WhatIsWrittenInAnEnsuresIsQuotedOverTheRowsTest {
     void aBehaviorWithNoRowToFillIsNotQuoted() {
         // The composition and not what is offered: this filling is built here rather than searched
         // for, so there is no store to ask what its rows would settle.
-        String block = GeneratedRows.of(souther.compiler.query.Offering.composed(
+        String block = GeneratedRows.of(souther.compiler.query.Composition.composed(
                         souther.compiler.query.OfferingRequest.overTheModule("example.todo", true),
-                        Map.of("findTodo", nothingOffered()), null),
+                        Map.of("findTodo", nothingOffered()), null).offeringEverything(),
                 Map.of("findTodo", List.of("ensures asked = NotFound -> id.value > 0")),
                 SourceNameResolver.identity()).text();
